# api/swagger/pt-BR.yaml
openapi: 3.0.3
info:
  title: Job Match API
  description: |
    ## API REST para avaliação de currículos e correspondência com vagas

    ### Changelog

    **Versão 0.1.0**
    - Implementado `POST /api/v1/file` para envio de arquivos DOC, DOCX e PDF que serão convertidos para texto no banco de dados
    - Implementado `POST /api/v1/user` para criação de usuários com autenticação
  version: 0.1.0
  contact:
    name: Bráulio Figueiredo
    email: braulio@braulioti.com.br
    url: http://brau.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://127.0.0.1:5000
    description: Servidor local de desenvolvimento
  - url: https://job-match-api.brau.io
    description: Servidor de produção

tags:
  - name: Health
    description: Endpoints de verificação de saúde da API
  - name: Files
    description: Endpoints para gerenciamento de arquivos
  - name: Users
    description: Endpoints para gerenciamento de usuários

paths:
  /health:
    get:
      tags:
        - Health
      summary: Verifica o status da API
      description: Retorna o status de saúde da API
      operationId: healthCheck
      responses:
        '200':
          description: API está funcionando corretamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy
                  service:
                    type: string
                    example: Job Match API

  /api/v1:
    get:
      tags:
        - Health
      summary: Informações da API
      description: Retorna informações sobre a API e seus endpoints
      operationId: apiInfo
      responses:
        '200':
          description: Informações da API
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Job Match API
                  version:
                    type: string
                    example: 1.0.0
                  endpoints:
                    type: object
                    properties:
                      health:
                        type: string
                        example: /health
  /api/v1/file:
    post:
      tags:
        - Files
      summary: Faz upload de um arquivo
      description: |
        Faz upload de um arquivo e armazena seus metadados no banco de dados PostgreSQL.
        
        O arquivo será processado para:
        - Gerar um UUID v4 único para o arquivo
        - Extrair metadados (nome original, extensão)
        - Converter e extrair conteúdo de texto para arquivos DOC e PDF
        
        **Formatos suportados:**
        - Conversão automática para MD: .doc, .docx e .pdf
        - Somente formatos que possam ser convertidos serão válidos
        
        **Validações:**
        - Arquivo não pode estar vazio
        - Arquivo deve ser fornecido no campo 'file' do formulário
      operationId: uploadFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Arquivo a ser enviado (multipart/form-data)
            encoding:
              file:
                contentType: application/octet-stream
      responses:
        '201':
          description: Arquivo enviado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
              examples:
                success:
                  summary: Upload bem-sucedido
                  value:
                    original_file_name: "curriculum.pdf"
                    hash: "a6d2335e-3536-4397-8635-c94de582eef3"
                    created_at: "2024-01-15T10:30:00Z"
                    updated_at: "2024-01-15T10:30:00Z"
        '400':
          description: Erro de validação (arquivo vazio ou não fornecido)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                noFile:
                  summary: Nenhum arquivo fornecido
                  value:
                    error: "Bad request"
                    message: "No file provided"
                emptyFile:
                  summary: Arquivo vazio
                  value:
                    error: "Bad request"
                    message: "File is empty"
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                serverError:
                  summary: Erro interno
                  value:
                    error: "Internal server error"
                    message: "Failed to save file to database: ..."

  /api/v1/user:
    post:
      tags:
        - Users
      summary: Cria um novo usuário
      description: |
        Cria um novo usuário no sistema com email e senha.
        
        O usuário será processado para:
        - Validar formato de email
        - Validar senha (mínimo 6 caracteres)
        - Gerar hash da senha usando werkzeug
        - Gerar UUID v4 único para autenticação
        
        **Validações:**
        - Email é obrigatório e deve ter formato válido
        - Senha é obrigatória e deve ter no mínimo 6 caracteres
        - Email deve ser único (não pode já existir no sistema)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
            examples:
              validUser:
                summary: Usuário válido
                value:
                  email: "usuario@example.com"
                  password: "senha123"
      responses:
        '201':
          description: Usuário criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserCreateResponse'
              examples:
                success:
                  summary: Usuário criado com sucesso
                  value:
                    id: 1
                    email: "usuario@example.com"
                    hash: "a6d2335e-3536-4397-8635-c94de582eef3"
                    created_at: "2024-01-15T10:30:00Z"
                    updated_at: "2024-01-15T10:30:00Z"
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidEmail:
                  summary: Email inválido
                  value:
                    error: "Validation error"
                    message: "Invalid email format"
                shortPassword:
                  summary: Senha muito curta
                  value:
                    error: "Validation error"
                    message: "Password must be at least 6 characters long"
                duplicateEmail:
                  summary: Email já existe
                  value:
                    error: "Validation error"
                    message: "User with this email already exists"
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                serverError:
                  summary: Erro interno
                  value:
                    error: "Internal server error"
                    message: "Failed to create user: ..."

components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Mensagem de erro
          example: Name is required
      required:
        - error
    
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Tipo do erro
          example: Validation error
        message:
          type: string
          description: Mensagem detalhada do erro
          example: job_vacancy cannot be empty
      required:
        - error
        - message

    FileUploadResponse:
      type: object
      properties:
        original_file_name:
          type: string
          description: Nome original do arquivo enviado
          example: curriculum.pdf
        hash:
          type: string
          description: Hash UUID v4 gerado para o arquivo
          example: a6d2335e-3536-4397-8635-c94de582eef3
        created_at:
          type: string
          format: date-time
          nullable: true
          description: Data e hora de criação do registro
          example: "2024-01-15T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          nullable: true
          description: Data e hora da última atualização do registro
          example: "2024-01-15T10:30:00Z"
      required:
        - original_file_name
        - hash
        - created_at
        - updated_at

    UserCreateRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Endereço de email do usuário
          example: "usuario@example.com"
        password:
          type: string
          format: password
          description: Senha do usuário (mínimo 6 caracteres)
          example: "senha123"
      required:
        - email
        - password

    UserCreateResponse:
      type: object
      properties:
        email:
          type: string
          format: email
          description: Endereço de email do usuário
          example: "usuario@example.com"
        hash:
          type: string
          description: Hash UUID v4 gerado para autenticação do usuário
          example: "a6d2335e-3536-4397-8635-c94de582eef3"
      required:
        - email
        - hash

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Autenticação via token JWT (para implementação futura)